refactor(header): move SettingsPanelContent to SettingsPanelContent.tsx and ThemeSettingsPopover to ThemeSettingsPopover.tsx
