+++ Frontend/src/components/Header/Header.tsx
@@ -2,7 +2,6 @@ import React, { useEffect, useState } from 'react'
 import { Link, useLocation, useNavigate } from 'react-router-dom'
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
-import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover'
 import { cn } from '@/lib/utils'
 import { useAppSelector } from '@/store/hooks'
 import Container from '../Container/Container'
@@ -12,7 +11,8 @@ import UserAvatar from '../User/UserAvatar'
 import { normalizeBrandHex, useTheme } from '../ThemeProvider'
 import { motion, useScroll, useMotionValueEvent } from 'framer-motion'
 import { Accordion, AccordionItem, AccordionTrigger, AccordionContent } from '@/components/ui/accordion'
-
+import ThemeSettingsPopover from '@/components/Theme/ThemeSettingsPopover'
+import SettingsPanelContent from '@/components/Theme/SettingsPanelContent'
 const navItems = [
   { name: 'Home', slug: '/' }
 ]
@@ -30,125 +30,12 @@ const headerVariants = {
   },
 };
 
-type ThemeMode = 'light' | 'dark' | 'system'
 
-// Extracted Settings Content
-function SettingsPanelContent({
-  theme,
-  setTheme,
-  brandColor,
-  setBrandColor,
-  resetBrandColor,
-  colorInput,
-  setColorInput,
-  colorError,
-  setColorError,
-  commitBrandColor,
-}: {
-  theme: ThemeMode
-  setTheme: (mode: ThemeMode) => void
-  brandColor: string
-  setBrandColor: (color: string) => void
-  resetBrandColor: () => void
-  colorInput: string
-  setColorInput: (value: string) => void
-  colorError: string
-  setColorError: (value: string) => void
-  commitBrandColor: () => void
-}) {
-  return (
-    <>
-      <div className="mb-4">
-        <span className="block text-xs font-medium text-on-surface-variant mb-2">Appearance</span>
-        <div className="grid grid-cols-3 gap-2">
-          {(['light', 'dark', 'system'] as const).map((m) => (
-            <button
-              key={m}
-              onClick={() => setTheme(m)}
-              className={`h-8 text-[11px] font-bold uppercase tracking-wider rounded-control border transition-all cursor-pointer ${theme === m
-                ? 'bg-primary text-on-primary border-transparent shadow-sm'
-                : 'border-outline-variant text-on-surface-variant hover:bg-surface-container hover:text-on-surface'
-                }`}
-            >
-              {m}
-            </button>
-          ))}
-        </div>
-      </div>
 
-      <div>
-        <span className="block text-xs font-medium text-on-surface-variant mb-2">Accent Color</span>
-        <div className="flex items-center gap-3">
-          <div className="relative size-8 rounded-full border border-outline-variant shadow-inner transition-colors duration-300 shrink-0 overflow-hidden">
-            <input
-              type="color"
-              value={brandColor}
-              onChange={(e) => {
-                setColorInput(e.target.value);
-                setColorError('');
-                setBrandColor(e.target.value);
-              }}
-              className="absolute inset-0 size-full scale-150 cursor-pointer opacity-0"
-              aria-label="Accent color picker"
-            />
-            <div className="size-full" style={{ backgroundColor: brandColor }} />
-          </div>
-          <div className="relative flex-1">
-            <input
-              type="text"
-              value={colorInput}
-              onChange={(e) => {
-                setColorInput(e.target.value);
-                setColorError('');
-              }}
-              onBlur={commitBrandColor}
-              onKeyDown={(e) => {
-                if (e.key === 'Enter') commitBrandColor();
-              }}
-              placeholder="#769CDF"
-              aria-invalid={Boolean(colorError)}
-              className="h-8 w-full rounded-control border border-outline-variant bg-surface-container px-3 text-xs font-mono text-on-surface outline-none focus:border-primary aria-invalid:border-error"
-            />
-          </div>
-        </div>
-        {colorError ? (
-          <p className="mt-2 text-xs font-medium text-error">{colorError}</p>
-        ) : null}
-        <button
-          type="button"
-          onClick={resetBrandColor}
-          className="mt-3 h-8 w-full rounded-control border border-outline-variant px-control-x text-xs font-semibold text-on-surface-variant transition-colors hover:border-outline hover:bg-surface-container hover:text-on-surface"
-        >
-          Reset Accent
-        </button>
-      </div>
-    </>
-  )
-}
 
-type ThemeSettingsPopoverProps = {
-  open: boolean
-  onOpenChange: (open: boolean) => void
-  trigger: React.ReactNode   // the button that goes in PopoverTrigger
-  children: React.ReactNode  // whatever goes inside PopoverContent
-  sideOffset?: number
-  options?: string
-}
 
-const settingsPanelClassName = "w-[min(18rem,calc(100vw-2rem))] rounded-card border border-outline-variant bg-surface-container-highest shadow-elevation-2"
 
-function ThemeSettingsPopover({ open, onOpenChange, trigger, children, sideOffset = 8, options}: ThemeSettingsPopoverProps) {
-  return (
-    <Popover open={open} onOpenChange={onOpenChange}>
-      <PopoverTrigger asChild>
-        {trigger}
-      </PopoverTrigger>
-      <PopoverContent align="end" sideOffset={sideOffset} className={cn(settingsPanelClassName, options)}>
-        {children}
-      </PopoverContent>
-    </Popover>
-  )
-}
+
 
 function Header() {
   const authStatus = useAppSelector((state) => state.auth.status)
@@ -157,23 +44,23 @@ function Header() {
   const navigate = useNavigate()
   const [searchValue, setSearchValue] = useState('')
   const [isMobileNavOpen, setIsMobileNavOpen] = useState(false)
-
+  
   const [isAuthSettingsOpen, setIsAuthSettingsOpen] = useState(false)
   const [isDesktopSettingsOpen, setIsDesktopSettingsOpen] = useState(false)
-
+  
   const isAuthRoute = location.pathname === '/login' || location.pathname === '/signup'
   const navControlClass = 'h-control-h'
   const userRole = useAppSelector((state) => state.auth.userData?.role);
   const TOP_THRESHOLD = 80;
-
+  
   const { theme, setTheme, brandColor, setBrandColor, resetBrandColor } = useTheme();
   const [colorInput, setColorInput] = useState(brandColor);
   const [colorError, setColorError] = useState('');
-
+  
   const { scrollY } = useScroll();
   const [headerHidden, setHeaderHidden] = useState(false);
   const headerHiddenRef = React.useRef(false);
-
+  
   useMotionValueEvent(scrollY, "change", (latest) => {
     const previous = scrollY.getPrevious() ?? 0;
     if (latest < TOP_THRESHOLD) {
@@ -197,39 +84,39 @@ function Header() {
       }
     }
   });
-
+  
   useEffect(() => {
     setColorInput(brandColor);
     setColorError('');
   }, [brandColor]);
-
+  
   const navRoutes = navItems;
 
   useEffect(() => {
     const params = new URLSearchParams(location.search)
     setSearchValue(params.get('q') ?? '')
   }, [location.search])
-
+  
   useEffect(() => {
     setIsMobileNavOpen(false)
   }, [location.pathname, location.search])
-
+  
   const handleSearch = (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault()
     const nextQuery = searchValue.trim()
     const params = new URLSearchParams(location.search)
     if (nextQuery) params.set('q', nextQuery)
-    else params.delete('q')
-
+      else params.delete('q')
+    
     const pathname = location.pathname === '/' || location.pathname === '/all-posts'
-      ? location.pathname : '/all-posts'
-
+    ? location.pathname : '/all-posts'
+    
     navigate({
       pathname,
       search: params.toString() ? `?${params.toString()}` : '',
     })
   }
-
+  
   const commitBrandColor = () => {
     const normalizedColor = normalizeBrandHex(colorInput);
     if (!normalizedColor) {
@@ -240,7 +127,7 @@ function Header() {
     setColorError('');
     setBrandColor(normalizedColor);
   }
-
+  
   const settingsPanelProps = { theme, setTheme, brandColor, setBrandColor, resetBrandColor, colorInput, setColorInput, colorError, setColorError, commitBrandColor }
 
   // --- Auth Route Header ---
